+++ custom_components/dessmonitor/device_support/device_registry.py
@@ -56,6 +56,10 @@ def _load_device_configurations() -> None:
 
         _register_devcode(6544, config_6544)
 
+        from .devcode_6514 import DEVCODE_CONFIG as config_6514
+
+        _register_devcode(6514, config_6514)
+
         from .devcode_6515 import DEVCODE_CONFIG as config_6515
 
         _register_devcode(6515, config_6515)